Refactor JNI for VAD
