+++ android/SherpaOnnxVad/app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:theme="@style/Theme.SherpaOnnxVad"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.k2fsa.sherpa.onnx.vad.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
