+++ sherpa-onnx/jni/CMakeLists.txt
@@ -17,6 +17,7 @@ set(sources
   speaker-embedding-extractor.cc
   speaker-embedding-manager.cc
   spoken-language-identification.cc
+  voice-activity-detector.cc
 )
 
 if(SHERPA_ONNX_ENABLE_TTS)
